move GenDocFromAutomodsumm from automodsumm.py to generate.py
